requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
pydantic==2.10.3
pydantic_core==2.27.1
python-multipart==0.0.19
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
zipp==3.20.2
